adapt from numpy 1.x to numpy 2.x

+++ pyvistools/oviseval.py
@@ -166,8 +166,8 @@ class OVISeval:
 
         evaluateVid = self.evaluateVid
         maxDet = p.maxDets[-1]
-        
-        
+
+
         self.evalImgs = [evaluateVid(vidId, catId, areaRng, occRng, maxDet)
                  for catId in catIds
                  for areaRng in p.areaRng
@@ -419,8 +419,8 @@ class OVISeval:
                         tps = np.logical_and(               dtm,  np.logical_not(dtIg) )
                         fps = np.logical_and(np.logical_not(dtm), np.logical_not(dtIg) )
 
-                        tp_sum = np.cumsum(tps, axis=1).astype(dtype=np.float)
-                        fp_sum = np.cumsum(fps, axis=1).astype(dtype=np.float)
+                        tp_sum = np.cumsum(tps, axis=1).astype(dtype=np.float64)
+                        fp_sum = np.cumsum(fps, axis=1).astype(dtype=np.float64)
                         for t, (tp, fp) in enumerate(zip(tp_sum, fp_sum)):
                             tp = np.array(tp)
                             fp = np.array(fp)
@@ -567,8 +567,8 @@ class Params:
         self.vidIds = []
         self.catIds = []
         # np.arange causes trouble.  the data point on arange is slightly larger than the true value
-        self.iouThrs = np.linspace(.5, 0.95, np.round((0.95 - .5) / .05) + 1, endpoint=True)
-        self.recThrs = np.linspace(.0, 1.00, np.round((1.00 - .0) / .01) + 1, endpoint=True)
+        self.iouThrs = np.linspace(.5, 0.95, int(np.round((0.95 - .5) / .05)) + 1, endpoint=True)
+        self.recThrs = np.linspace(.0, 1.00, int(np.round((1.00 - .0) / .01)) + 1, endpoint=True)
         self.maxDets = [20]
         self.areaRng = [[0 ** 2, 1e5 ** 2], [32 ** 2, 96 ** 2], [96 ** 2, 1e5 ** 2]]
         self.areaRngLbl = ['all', 'medium', 'large']
